scripts/requirements.txt
python-dotenv==1.0.1
typesense==0.21.0
